Update User Login Logout Functionality
